Replaced_and_header
The header and footer with Bootstrap React were embedded in the history page.
In the footer, the fixed bottom in className also changed because it makes overlapped with right side textboxes from the bottom
