--- client/src/Components/History/Footer.css
@@ -1,11 +1,0 @@
-.footer{
-    background-color: hsl(360, 83%, 60%);
-    padding: 10px;
-    display: flex;
-    justify-content: flex-end;
-}
-
-.footer img {
-    width: 160px;
-
-}
