--- client/src/Components/History/Footer.js
@@ -1,12 +1,0 @@
-import React from "react";
-import AryaLogo from "../../assets/TeamLogo.png";
-import "./Footer.css";
-const Footer = () => {
-	return (
-		<div className="footer">
-			<img src={AryaLogo} alt="logo" />
-		</div>
-	);
-};
-
-export default Footer;
