--- client/src/Components/History/Header.css
@@ -1,79 +1,0 @@
-* {
-	box-sizing: border-box;
-}
-
-button {
-	border: transparent;
-	background: none;
-}
-
-.navbar {
-	background-color: #f2eff0;
-	padding: 16px;
-	display: flex;
-	align-items: center;
-	justify-content: space-between;
-}
-
-.navbar img {
-	width: 120px;
-}
-
-.navbar h1 {
-	color: #ee4344;
-	padding: 10px 56px 0 0;
-	font-size: 16px;
-	font-weight: 900;
-}
-
-.singout-container button {
-	display: flex;
-	align-items: center;
-	justify-content: center;
-	flex-direction: column;
-	width: 40px;
-	font-size: 40px;
-}
-
-.singout-container p {
-	margin: 0;
-	font-size: 10px;
-}
-
-.singout-container svg {
-	padding: 0;
-	margin: 0;
-}
-
-.header-red-div {
-	padding: 16px;
-	background-color: #ee4344;
-}
-
-.header-red-div h5 {
-	padding-top: 4px;
-	color: #fff;
-	font-weight: 700;
-}
-
-.slogan-container {
-	display: flex;
-	justify-content: space-between;
-	align-items: center;
-	gap: 16px;
-}
-
-.slogan-container button {
-	font-size: 22px;
-	color: #fff;
-	width: 120px;
-	font-weight: 600;
-	background: rgb(142, 8, 8);
-	border-radius: 5px;
-}
-
-@media only screen and (min-width: 700px) {
-	.navbar h1 {
-		font-size: 28px;
-	}
-}
