+++ client/src/pages/Main.css
@@ -1,3 +1,8 @@
+button {
+	border: none;
+	background: none;
+}
+
 .search-input-form {
 	display: flex;
 	flex-direction: column;
